move router from server to a dependent crate.
