+++ router/src/hash.rs
