+++ server/src/lib.rs
@@ -8,7 +8,6 @@
 #[macro_use]
 extern crate common_util;
 
-mod avro_util;
 pub mod config;
 mod consts;
 mod context;
@@ -22,7 +21,6 @@ pub mod local_tables;
 pub mod logger;
 mod metrics;
 mod mysql;
-pub mod route;
 pub mod schema_config_provider;
 pub mod server;
 pub mod table_engine;
